Fix -Wsign-conversion warnings in sparse_matvec_block size_t casts

+++ src/sparse_matrix.c
@@ -781,7 +781,7 @@ sparse_err_t sparse_matvec_block(const SparseMatrix *mat, const double *X, idx_t
             double a_ij = node->value;
             for (idx_t k = 0; k < nrhs; k++) {
                 size_t ok = (size_t)m * (size_t)k;
-                Y[log_i + ok] += a_ij * X[log_j + (size_t)mat->cols * (size_t)k];
+                Y[(size_t)log_i + ok] += a_ij * X[(size_t)log_j + (size_t)mat->cols * (size_t)k];
             }
             node = node->right;
         }
